backend/requirements.txt
# Core ML dependencies
tensorflow>=2.10.0
numpy>=1.21.0
scipy>=1.7.0

# Image processing
opencv-python>=4.5.0
Pillow>=9.0.0

# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Utilities
scikit-learn>=1.0.0
matplotlib>=3.5.0

# CORS support
python-dotenv>=1.0.0

